Heavy refactoring, introduce DCAUtils
Separate most utility functions in a utils module,
intended to become an autonomouse package eventually.
